+++ TabletopTweaks/Extensions/ExtentionMethods.cs
@@ -149,8 +149,12 @@ namespace TabletopTweaks.Extensions {
                 if (selection.m_AllFeatures.Contains(featureReference)) {
                     selection.m_AllFeatures = selection.m_AllFeatures.Where(f => !f.Equals(featureReference)).ToArray();
                 }
+                if (selection.m_Features.Contains(featureReference)) {
+                    selection.m_Features = selection.m_Features.Where(f => !f.Equals(featureReference)).ToArray();
+                }
             }
             selection.m_AllFeatures = selection.m_AllFeatures.OrderBy(feature => feature.Get().Name).ToArray();
+            selection.m_Features = selection.m_Features.OrderBy(feature => feature.Get().Name).ToArray();
         }
 
         public static void AddFeatures(this BlueprintFeatureSelection selection, params BlueprintFeature[] features) {
@@ -159,8 +163,12 @@ namespace TabletopTweaks.Extensions {
                 if (!selection.m_AllFeatures.Contains(featureReference)) {
                     selection.m_AllFeatures = selection.m_AllFeatures.AppendToArray(featureReference);
                 }
+                if (!selection.m_Features.Contains(featureReference)) {
+                    selection.m_Features = selection.m_Features.AppendToArray(featureReference);
+                }
             }
             selection.m_AllFeatures = selection.m_AllFeatures.OrderBy(feature => feature.Get().Name).ToArray();
+            selection.m_Features = selection.m_Features.OrderBy(feature => feature.Get().Name).ToArray();
         }
         public static void AddPrerequisiteFeature(this BlueprintFeature obj, BlueprintFeature feature) {
             obj.AddPrerequisiteFeature(feature, GroupType.All);
@@ -217,6 +225,12 @@ namespace TabletopTweaks.Extensions {
                     break;
                 default:
                     break;
+            }
+        }
+
+        public static void AddPrerequisites<T>(this BlueprintFeature obj, params T[] prerequisites) where T : Prerequisite {
+            foreach (var prerequisite in prerequisites) {
+                obj.AddPrerequisite(prerequisite);
             }
         }